+++ pkgs/intl_translation/test/message_extraction/print_to_list.dart
@@ -1,3 +1,7 @@
+// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 /// This provides a way for a test to print to an internal list so the
 /// results can be verified rather than writing to and reading a file.