+++ pkgs/intl4x/lib/src/datetime_format/datetime_format_impl.dart
@@ -3,7 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import '../ecma/ecma_policy.dart';
-import '../locale.dart';
+import '../locale/locale.dart';
 import '../options.dart';
 import '../utils.dart';
 import 'datetime_format_4x.dart';