+++ pkgs/intl4x/lib/src/number_format/number_format_ecma.dart
@@ -5,7 +5,7 @@
 import 'package:js/js.dart';
 import 'package:js/js_util.dart';
 
-import '../locale.dart';
+import '../locale/locale.dart';
 import '../options.dart';
 import 'number_format_impl.dart';
 import 'number_format_options.dart';