+++ pkgs/intl4x/lib/src/collation/collation_ecma.dart
@@ -5,7 +5,7 @@
 import 'package:js/js.dart';
 import 'package:js/js_util.dart';
 
-import '../locale.dart';
+import '../locale/locale.dart';
 import '../options.dart';
 import 'collation_impl.dart';
 import 'collation_options.dart';