+++ pkgs/intl4x/lib/src/display_names/display_names_ecma.dart
@@ -5,7 +5,7 @@
 import 'package:js/js.dart';
 import 'package:js/js_util.dart';
 
-import '../locale.dart';
+import '../locale/locale.dart';
 import '../options.dart';
 import 'display_names_impl.dart';
 import 'display_names_options.dart';