+++ pkgs/intl/CHANGELOG.md
@@ -3,6 +3,7 @@
  * Add `tryParse` to NumberFormat.
  * Add `tryParse` `tryParseStrict` `tryParseLoose` `tryParseUtc` to DateFormat.
  * Add fallback for deprecated locales, such as `he` <-> `iw`.
+ * Update to CLDR v43.
 
 ## 0.18.1
  * Update ruble sign and update corresponding test.