+++ pkgs/intl/CHANGELOG.md
@@ -5,6 +5,7 @@
  * Add fallback for deprecated locales, such as `he` <-> `iw`.
  * Update to CLDR v43.
  * Switch QAR currency name to Riyal.
+ * Fix standalone month pattern mapping for `el` locale 
 
 ## 0.18.1
  * Update ruble sign and update corresponding test.