+++ pkgs/intl/CHANGELOG.md
@@ -11,6 +11,7 @@
  * Switch to `3.0.0` SDK.
  * Fix issue [#483](https://github.com/dart-lang/i18n/issues/483) about date parsing with a `yy` skeleton.
  * Extract `NumberParserBase` abstract class
+ * Update to CLDR v44.
 
 ## 0.18.1
  * Update ruble sign and update corresponding test.